SPADE/requirements.txt
spade==3.1.8
fastapi==0.68.0
uvicorn==0.15.0
paho-mqtt==1.6.1
pydantic==1.8.2
requests==2.28.1